adding to order work incorrect

+++ core/src/main/java/com/shop/itis/domain/Order.java
@@ -2,7 +2,9 @@ package com.shop.itis.domain;
 
 
 import javax.persistence.*;
-import java.util.*;
+import java.util.Date;
+import java.util.HashSet;
+import java.util.Set;
 
 
 @Entity
@@ -11,7 +13,7 @@ public class Order {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     Address address;
@@ -47,11 +49,11 @@ public class Order {
         this.pay_type = pay_type;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
